+++ src/ByteDev.Common/Time/IDateTimeProvider.cs
@@ -5,21 +5,21 @@ namespace ByteDev.Common.Time
     /// <summary>
     /// Provides a way to get <see cref="T:System.DateTime" /> through an interface.
     /// </summary>
-    public class DateTimeProvider : IDateTimeProvider
+    public interface IDateTimeProvider
     {
         /// <summary>
         /// Returns <see cref="T:System.DateTime" />.Now.
         /// </summary>
-        public DateTime Now => DateTime.Now;
+        DateTime Now { get; }
 
         /// <summary>
         /// Returns <see cref="T:System.DateTime" />.Today.
         /// </summary>
-        public DateTime Today => DateTime.Today;
+        DateTime Today { get; }
 
         /// <summary>
         /// Returns <see cref="T:System.DateTime" />.UtcNow.
         /// </summary>
-        public DateTime UtcNow => DateTime.UtcNow;
+        DateTime UtcNow { get; }
     }
 }