+++ src/ByteDev.Common/Time/TimespanExtensions.cs
@@ -16,8 +16,10 @@ namespace ByteDev.Common.Time
         {
             if (source.Days > 0)
                 return source.Days + " day".Pluralize(source.Days);
+
             if (source.Hours > 0)
                 return source.Hours + " hour".Pluralize(source.Hours);
+
             if (source.Minutes > 0)
                 return source.Minutes + " minute".Pluralize(source.Minutes);